Refactoring of PhotonVisibilityExport_module
The different visibility models are now implemented as swappable larsoft
tools. This commit also removes a deprecated dependency on Legacy LArG4.
