+++ duneopdet/PhotonPropagation/NeedsTensorflow/PhotonVisibilityExport.fcl
@@ -7,6 +7,7 @@ semianalytical_model_settings: {
   do_include_anode_refl:    false
   vuvhitspars:              @local::dunefd_pdfastsim_par_ar.VUVHits
   vishitspars:              @local::dunefd_pdfastsim_par_ar_refl.VISHits
+  OpticalPathTool:          @local::dunefd_pdfastsim_par_ar.OpticalPathTool
 }
 
 computablegraph_settings: {
@@ -27,32 +28,33 @@ standard_photovisexport:
    n_vis_samplings:               "5"     # nr of visibility samplings inside the voxel (min 1)
 
    tpc_vis_model: {
-      model_type: "" # visibility model to be used
+      tool_type: "" # visibility model to be used
       settings: {}
    }      
 
    do_include_buffer: true   # include buffer region 
    buf_vis_model: {
-    model_type: ""
+    tool_type: ""
     settings: {}
    }
 }
 
 photovisexport_fdhd: @local::standard_photovisexport
-photovisexport_fdhd.tpc_vis_model.model_type: "semianalytical"
+photovisexport_fdhd.tpc_vis_model.tool_type: "SemiAnalyticalVisModel"
 photovisexport_fdhd.tpc_vis_model.settings: @local::semianalytical_model_settings
-photovisexport_fdhd.buf_vis_model.model_type: "photonlibrary"
+photovisexport_fdhd.buf_vis_model.tool_type: "PhotonLibraryVisModel"
 photovisexport_fdhd.buf_vis_model.settings: {}
 
 photovisexport_fdvd_ar: @local::photovisexport_fdhd
 photovisexport_fdvd_ar.tpc_vis_model.settings.vuvhitspars: @local::dunevd_pdfastsim_par_ar.VUVHits
 photovisexport_fdvd_ar.tpc_vis_model.settings.do_refl: @local::dunevd_pdfastsim_par_ar.DoReflectedLight
+photovisexport_fdvd_ar.tpc_vis_model.settings.OpticalPathTool: @local::dunevd_pdfastsim_par_ar.OpticalPathTool
 photovisexport_fdvd_ar.tpc_vis_model.settings.do_include_anode_refl: @local::dunevd_pdfastsim_par_ar.IncludeAnodeReflections
 
 photovisexport_fdvd_xe: @local::photovisexport_fdhd
 photovisexport_fdvd_xe.tpc_vis_model.settings.vuvhitspars: @local::dunevd_pdfastsim_par_xe.VUVHits
 photovisexport_fdvd_xe.tpc_vis_model.settings.do_refl: @local::dunevd_pdfastsim_par_xe.DoReflectedLight
 photovisexport_fdvd_xe.tpc_vis_model.settings.do_include_anode_refl: @local::dunevd_pdfastsim_par_xe.IncludeAnodeReflections
-
+photovisexport_fdvd_xe.tpc_vis_model.settings.OpticalPathTool: @local::dunevd_pdfastsim_par_xe.OpticalPathTool
 END_PROLOG
