+++ duneopdet/PhotonPropagation/NeedsTensorflow/VisibilityModelTools/ComputableGraphVisModel_tool.cc
@@ -1,0 +1,4 @@
+#include "art/Utilities/ToolMacros.h"
+#include "duneopdet/PhotonPropagation/NeedsTensorflow/VisibilityModelTools/ComputableGraphVisModel.hh"
+
+DEFINE_ART_CLASS_TOOL(phot::vismodel::ComputableGraphVisModel)
