+++ duneopdet/PhotonPropagation/NeedsTensorflow/VisibilityModelTools/PhotonLibraryVisModel_tool.cc
@@ -1,0 +1,5 @@
+#include "art/Utilities/ToolMacros.h"
+#include "duneopdet/PhotonPropagation/NeedsTensorflow/VisibilityModelTools/PhotonLibraryVisModel.hh"
+
+DEFINE_ART_CLASS_TOOL(phot::vismodel::PhotonLibraryVisModel)
+
